refactor: move mockNeighbors
